winex11: Copy icon color table in 256-color mode (Valgrind)
create_icon_pixmaps calls GetDIBits and create_pixmap_from_image, which
calls get_dib_info_size. GetDIBits sets biClrUsed to 0, which is a
special value meaning that the size of the color table is the maximum
size possible given the bit depth. Without code to handle that case in
get_dib_info_size, the color table is lost.

gdi32's version of get_dib_info_size does not have to deal with
biClrUsed being zero because bitmapinfo_from_user_bitmapinfo makes a
copy of the bitmap info with biClrUsed set to a nonzero value. However,
there is no reason to have anything that complicated in winex11.

Signed-off-by: Alex Henrie <alexhenrie24@gmail.com>